Implement a clear class hierarchy
This commit introduces an ExpertBot class that all expert bots should
inherit from. To have a central place for storing the four types of bots
there is now a datatypes module that contains an enum listing the four
bot types. It can and should be used for more types that are part of the
IntelMQ codebase.
Also some output and parser bots were updated to depend on the correct
bot type.